refactor(@theme-ui/core): stop depending on styled-system
